update geary hook
