+++ common/build.gradle.kts
@@ -9,11 +9,10 @@ dependencies {
     compileOnly("org.jetbrains:annotations:23.0.0")
     compileOnly("com.comphenix.protocol:ProtocolLib:5.0.0")
     compileOnly("me.clip:placeholderapi:2.11.1")
-    compileOnly("com.ticxo.modelengine:api:R3.0.1")
+    compileOnly("com.ticxo:modelengine:R3.0.1")
     compileOnly("com.github.oraxen:oraxen:-SNAPSHOT")
     compileOnly("com.github.LoneDev6:API-ItemsAdder:3.2.5")
-    compileOnly("com.mineinabyss:geary-papermc-core:0.19.113")
-    compileOnly("com.mineinabyss:looty:0.8.67")
+    compileOnly("com.mineinabyss:geary-papermc:0.24-SNAPSHOT")
     compileOnly("com.hibiscus:hmccolor:0.3-SNAPSHOT")
     compileOnly("com.sk89q.worldguard:worldguard-bukkit:7.1.0-SNAPSHOT")
     compileOnly("it.unimi.dsi:fastutil:8.5.11")
